requirements.txt
torch==2.5.0
torchvision==0.20.0
accelerate
lightning~=2.1.2
pytorch_lightning
torchdiffeq
torch-fidelity
torchmetrics[image]
absl-py
av==13.1.0
blobfile
click
colorama
decord
einops
fairscale
gluonts[torch]==0.13.1
h5py
hydra-core~=1.3.2
imageio
internetarchive
loguru
lpips
matplotlib
ml_collections
moviepy
omegaconf~=2.3.0
open_clip-torch
opencv-python
pandas
pyrealsense2
pytorchvideo~=0.1.5
pyzmq
rotary_embedding_torch
scikit-learn
timm
tqdm
transformers
wandb~=0.17.0
wandb-osh==1.2.1
webdataset
xformers
diffusers
